Python
Coded a reverse guessing game, where the user thinks of a number between 1 and 100. The computer takes guesses, and after each guess, the user is asked to state whether the guess is "too low", or "too high". Statistically if the code is in it's most efficient form the computer should take no more than 8 guesses to get the correct number. The code above fulfills said requirement.